fix: 코드리뷰 반영

+++ src/main/java/com/cookeep/cookeep/api/dto/CustomIngredientCreateResponseDto.java
@@ -10,7 +10,8 @@ import lombok.Getter;
 @AllArgsConstructor
 public class CustomIngredientCreateResponseDto {
 
-    private int customIngredientId;
+
+    private Long customIngredientId;
     private String name;
     private Integer expirationDays;
     private Storage storage;
